backend/requirements.txt
# Python 依赖清单
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
langchain==1.2.10
langchain-community==0.4.1
chromadb==1.5.2
psycopg2-binary==2.9.11
sqlalchemy==2.0.48
openai==2.26.0
opendataloader-pdf==2.2.0
python-multipart==0.0.22
sentence-transformers==5.3.0
python-docx==1.2.0